+++ process/neutronics.py
@@ -6,22 +6,11 @@ ISBN:9780471223634.
 import functools
 
 import numpy as np
-from neutronics_data import (
-    ZeroContinuousFunc,
-    breeding_xs_data_bank,
-    extract_atomic_mass,
-    get_avg_atomic_mass,
-    material_composition_data_bank,
-    material_density_data_bank,
-    xs_data_bank,
-)
-from numpy import testing as npt
-from scipy.constants import Avogadro
+from neutronics_data import MaterialMacroInfo
+from numpy import typing as npt
 from scipy.special import expm1
 
-BARNS_CM2 = 1e-24
-N_A = Avogadro
-N2N_Q_VALUE = ...
+from process.exceptions import ProcessValidationError
 
 
 def groupwise(func):
@@ -123,226 +112,16 @@ def extrapolation_length(diffusion_coefficient: float) -> float:
     return 0.7104 * 3 * diffusion_coefficient
 
 
-def calculate_average_macro_xs(
-    composition: dict[str, float],
-    micro_xs: dict[str, float | npt.NDArray[np.float64]],
-    density: float,
-) -> float | npt.NDArray[np.float64]:
-    r"""
-    Calculate the macroscopic cross-section for a specific energy group and a specific
-    reaction (a scalar value), when given the microscopic cross-section values of its
-    components and its density.
-
-    Parameters
-    ----------
-    composition:
-        Fraction of each species of atoms of the medium. Does not have to be normalised.
-        Given in the format {'species': float(fraction)}.
-    micro_xs:
-        A dictionary of each species, and their microscopic cross-sections, given in
-        [barns].
-        Given in the format {'species': float(microscopic cross-section)}.
-    density:
-        Density of the medium, given in [g/cm^3]
-
-    Notes
-    -----
-    .. math::
-        \Sigma &= N_d\sigma
-        &= \frac{N_A}{A} \rho \sigma
-    """
-
-    total_fraction = sum(composition.values())
-    weighted_micro_xs, weighted_atomic_mass = [], []
-    if composition.keys() != micro_xs.keys():
-        raise KeyError(
-            "The two dictionaries 'composition' and 'micro_xs' must have matching keys."
-        )
-    for species, fraction in composition.items():
-        normalized_fraction = fraction / total_fraction
-        weighted_atomic_mass.append(
-            normalized_fraction * extract_atomic_mass(species)
-        )
-        weighted_micro_xs.append(normalized_fraction * micro_xs[species])
-    avg_sigma = np.sum(weighted_micro_xs, axis=0)
-    avg_mass_amu = sum(weighted_atomic_mass)
-    return (
-        (BARNS_CM2 * N_A) / avg_mass_amu * avg_sigma * density
-    )  # N_A/A * rho * sigma
-
-
-def discretize_xs(
-    continuous_xs, group_structure: list[float]
-) -> npt.NDArray[np.float64]:
-    """
-    Discretise a continuous cross-section function into a group structure of n discrete
-    flaots.
-
-    Parameters
-    ----------
-    continuous_xs:
-        continuous cross-section function to be discretized.
-    group_structure:
-        group structure of neutron energies, the n+1 energy bin boundaries for the n
-        neutron groups, in descending energies.
-
-    Returns
-    -------
-    :
-        microscopic cross-section values discretized according to the group structure.
-    """
-    return
-
-
-def scattering_weight_matrix(
-    group_structure: list[float], atomic_mass: float
-) -> npt.NDArray:
-    """
-    Parameters
-    ----------
-    group_structure:
-        the n+1 energy bin boundaries for the n neutron groups, in descending energies.
-    atomic_mass:
-        atomic mass of the medium
-
-    Returns
-    -------
-    :
-        A lower triangular matrix, where the i-th row contains the normalized weights for
-        scattering down from the i-th bin to the j-th bin. The main-diagonal contains
-        the self-scattering cross-section.
-        e.g. [2,1] would be the fraction of elastic scattering reactions that causes
-        neutrons from group 3 to scatter into group 2.
-        The upper triangle must be all zeros.
-        np.sum(axis=1) == np.ones(len(group_structure)-1).
-    """
-    return
-
-
-def expand_macro_neutron_multiplication_xs_into_matrix(
-    discrete_n2n_xs: npt.NDArray[np.float64],
-    group_structure: list[float],
-    q_value: float,
-) -> npt.NDArray:
-    """Instead of only having the macroscopic cross-section values for the (n,2n)
-    reaction for each group, calculate the macroscopic cross-section of neutron in the
-    [i]-th bin producing a neutron in the [j]-bin, recorded as the [i,j] element in the
-    matrix.
-
-    Parameters
-    ----------
-    discrete_n2n_xs:
-        The group-wise macroscopic cross-section for the n2n reaction.
-        A 1D numpy array, with len==number of neutron groups.
-    group_structure:
-        The n+1 energy bin boundaries for the n neutron groups, in descending energies.
-    q_value:
-        The difference in q-value.
-
-    Returns
-    -------
-    :
-        A macroscopic neutron multiplication cross-section matrix, such that each row
-        should sum to = 2 * discrete_n2n_xs (since two neutrons should be produced per
-        neutron consumed in the (n,2n) reaction).
-
-    Notes
-    -----
-    This is a three body problem. TODO: further investigation is needed to figure out
-    how to distribute the two outputted neutron's energies!
-    """
-
-
-def get_material_nuclear_data(
-    material: str, group_structure: list[float]
-) -> tuple[npt.NDArray[np.float64], npt.NDArray, float]:
-    """
-    The constants that is directly used.
-
-    Parameters
-    ----------
-    materials:
-        Which material that we want to get the data out of.
-    group_structure:
-        the n+1 energy bin boundaries for the n neutron groups, in descending energies.
-
-    Returns
-    -------
-    discrete_macro_total_xs:
-        All group-wise total cross-sections, given in ascending order.
-    macroscopic scattering cross-section matrix:
-        A lower-triangular matrix of total scattering cross-sections,
-        e.g. [2,1] would be the scattering cross-section from neutron group 3 to group 2.
-        The upper triangle MUST be zeros.
-    avg_atomic_mass:
-        average atomic mass, used for further calcuations
-
-    Notes
-    -----
-    When calculating the group-wise diffusion coefficients, the i-th group's scattering
-    cross-section :math:`\\Sigma_{s}` is the i-th element on the main diagonal of the
-    macroscopic scattering cross-section matrix.
-    """
-    density = material_density_data_bank[material]
-    composition = material_composition_data_bank[material]
-    avg_atomic_mass = get_avg_atomic_mass(composition)
-
-    # dicts of {"isotope": npt.NDArray[np.float64] 1D arrays}
-    micro_total_xs = {}
-    micro_scattering_xs = {}
-    micro_n2n_xs = {}
-    for species in composition:
-        total_xs_continuous, elastic_scattering_xs_continuous = xs_data_bank[
-            species
-        ]
-        n2n_xs_continuous = breeding_xs_data_bank.get(
-            species, ZeroContinuousFunc
-        )
-        micro_total_xs[species] = discretize_xs(
-            total_xs_continuous, group_structure
-        )
-        micro_scattering_xs[species] = discretize_xs(
-            elastic_scattering_xs_continuous, group_structure
-        )
-        micro_n2n_xs[species] = discretize_xs(
-            n2n_xs_continuous, group_structure
-        )
-
-    discrete_macro_total_xs = calculate_average_macro_xs(
-        composition, micro_total_xs, density
-    )
-    discrete_macro_scattering_xs = calculate_average_macro_xs(
-        composition, micro_scattering_xs, density
-    )
-    source_matrix = (
-        scattering_weight_matrix(group_structure, avg_atomic_mass).T
-        * discrete_macro_scattering_xs
-    ).T
-    for (
-        modified_composition_file,
-        modified_density,
-        q_value,
-    ) in n2n_susceptible_species:
-        source_matrix += expand_macro_neutron_multiplication_xs_into_matrix(
-            calculate_average_macro_xs(composition, micro_n2n_xs, density),
-            group_structure,
-            q_value,
-        )
-
-    return discrete_macro_total_xs, source_matrix, avg_atomic_mass
-
-
 class NeutronFluxProfile:
     """Neutron flux in the first wall or the blanket."""
 
     def __init__(
         self,
         flux: float,
-        fw_mat: str,
         x_fw: float,
-        bz_mat: str,
         x_bz: float,
-        n_groups: int = 1,
+        fw_mat: MaterialMacroInfo,
+        bz_mat: MaterialMacroInfo,
     ):
         """Initialize a particular FW-BZ geometry and neutron flux.
 
@@ -351,44 +130,55 @@ class NeutronFluxProfile:
         flux:
             Nuetron flux directly emitted by the plasma, incident on the first wall.
             unit: cm^-2 s^-1
-        fw_mat:
-            first wall material
         x_fw:
             thickness of the first wall [m]. It will be converted and stored in [cm].
-        bz_mat:
-            blanket material
         x_bz:
             thickness of the blanket [m]. It will be converted and stored in [cm].
+        fw_mat:
+            first wall material information
+        bz_mat:
+            blanket material information
+
+        Attributes
+        ----------
+        x_fw:
+            thickness of the first wall, converted from [m] into [cm].
+        x_bz:
+            thickness of the blanket, converted from [m] into [cm].
         n_groups:
-            number of groups used to approximate this.
+            number of groups in the group structure
+        group_structure:
+            energy bin edges, 1D array of len = n_groups+1
         """
         self.flux = flux  # flux incident on the first wall.
         if not (0 < x_fw < x_bz):
             raise ValueError(
                 f"Cannot construct a first-wall+blanket module where{x_fw=}, {x_bz=}."
             )
-        self.fw_mat = fw_mat
         self.x_fw = x_fw * 100
-        self.bz_mat = bz_mat
         self.x_bz = x_bz * 100
-        self.n_groups = n_groups
+        self.fw_mat = fw_mat
+        self.bz_mat = bz_mat
+        self.n_groups = self.fw_mat.n_groups
+        self.group_structure = self.fw_mat.group_structure
+        if not np.isclose(
+            self.fw_mat.group_structure, self.bz_mat.group_structure
+        ):
+            raise ProcessValidationError(
+                "The first-wall material info and breeding zone material info"
+                "must have the same group structure!"
+            )
 
-        # macroscopic cross-sections Sigma saved here. Not capitalized due to style
-        self.fw_sigma_t, self.fw_sigma_s, self.fw_A = (
-            get_material_nuclear_data(self.fw_mat, n_groups)
-        )
-        self.bz_sigma_t, self.bz_sigma_s, self.bz_A = (
-            get_material_nuclear_data(self.bz_mat, n_groups)
-        )
+        # macroscopic cross-sections Sigma saved here.
         # Dictionaries indexed by integers, so that we can create these values
-        # out of sequence without messing up the order.
+        # out of sequence.
         self.integration_constants = {}
-        self.l_fw_2, self.l_bz_2 = {}, {}
+        self.l_fw_2, self.l_bz_2 = {}, {}  # diffusion lengths squared
         self.extended_boundary = {}
 
-    def solve_one_group(self) -> None:
+    def solve_lowest_group(self) -> None:
         """
-        Solve the first-group's neutron diffusion equation.
+        Solve the highest-energy (lowest-lethargy)-group's neutron diffusion equation.
         Store the solved constants in self.extended_boundary[0], self.l_fw_2[0],
         self.l_bz_2[0], and self.integration_constants[0].
         """
@@ -396,14 +186,14 @@ class NeutronFluxProfile:
         if i in self.integration_constants:
             return  # skip if it has already been solved.
         self.l_fw_2[i], d_fw = get_diffusion_coefficient_and_length(
-            self.fw_sigma_t[i],
-            self.fw_sigma_s[i, i],
-            self.fw_A,
+            self.fw_mat.sigma_t[i],
+            self.fw_mat.sigma_s[i, i],
+            self.fw_mat.avg_atomic_mass,
         )
         self.l_bz_2[i], d_bz = get_diffusion_coefficient_and_length(
-            self.bz_sigma_t[i],
-            self.bz_sigma_s[i, i],
-            self.bz_A,
+            self.bz_mat.sigma_t[i],
+            self.bz_mat.sigma_s[i, i],
+            self.bz_mat.avg_atomic_mass,
         )
         l_fw = np.sqrt(abs(self.l_fw_2[i]))
         l_bz = np.sqrt(abs(self.l_bz_2[i]))
@@ -465,7 +255,7 @@ class NeutronFluxProfile:
                 f"n must be a positive integer between 1 and {self.n_groups}!"
             )
         if n == 1:
-            return self.solve_one_group()
+            return self.solve_lowest_group()
         for k in range(n - 1):
             if k not in self.integration_constants:
                 self.solve_group_n(k)
@@ -473,14 +263,14 @@ class NeutronFluxProfile:
         if i in self.integration_constants:
             return None  # skip if it has already been solved.
         self.l_fw_2[i], d_fw = get_diffusion_coefficient_and_length(
-            self.fw_sigma_t[i],
-            self.fw_sigma_s[i, i],
-            self.fw_A,
+            self.fw_mat.sigma_t[i],
+            self.fw_mat.sigma_s[i, i],
+            self.fw_mat.avg_atomic_mass,
         )
         self.l_bz_2[i], d_bz = get_diffusion_coefficient_and_length(
-            self.bz_sigma_t[i],
-            self.bz_sigma_s[i, i],
-            self.bz_A,
+            self.bz_mat.sigma_t[i],
+            self.bz_mat.sigma_s[i, i],
+            self.bz_mat.avg_atomic_mass,
         )
         l_fw = np.sqrt(abs(self.l_fw_2[i]))
         l_bz = np.sqrt(abs(self.l_bz_2[i]))
@@ -597,9 +387,9 @@ class NeutronFluxProfile:
         l_fw = np.sqrt(abs(self.l_fw_2[i]))
         c1, c2, c3, c4 = self.integration_constants[i]
         if reaction_type == "non-scattering":
-            sigma = self.fw_sigma_t[i] - self.fw_sigma_s[i, i]
+            sigma = self.fw_mat.sigma_t[i] - self.fw_mat.sigma_s[i, i]
         elif reaction_type == "total":
-            sigma = self.fw_sigma_t[i]
+            sigma = self.fw_mat.sigma_t[i]
         else:
             raise NotImplementedError(
                 f"Not yet implemented the reaction type {reaction_type}"
@@ -625,9 +415,9 @@ class NeutronFluxProfile:
         l_bz = np.sqrt(abs(self.l_bz_2[i]))
         c1, c2, c3, c4 = self.integration_constants[i]
         if reaction_type == "non-scattering":
-            sigma = self.bz_sigma_t[i] - self.bz_sigma_s[i, i]
+            sigma = self.bz_mat.sigma_t[i] - self.bz_mat.sigma_s[i, i]
         elif reaction_type == "total":
-            sigma = self.bz_sigma_t[i]
+            sigma = self.bz_mat.sigma_t[i]
         else:
             raise NotImplementedError(
                 f"Not yet implemented the reaction type {reaction_type}"
@@ -646,9 +436,9 @@ class NeutronFluxProfile:
         i = n - 1
         c1, c2, c3, c4 = self.integration_constants[i]
         l_bz_2, d_bz = get_diffusion_coefficient_and_length(
-            self.bz_sigma_t[i],
-            self.bz_sigma_s[i, i],
-            self.bz_A,
+            self.bz_mat.sigma_t[i],
+            self.bz_mat.sigma_s[i, i],
+            self.bz_mat.avg_atomic_mass,
         )
         l_bz = np.sqrt(abs(l_bz_2))
         return (
@@ -665,9 +455,9 @@ class NeutronFluxProfile:
         i = n - 1
         c1, c2, c3, c4 = self.integration_constants[i]
         l_bz_2, d_bz = get_diffusion_coefficient_and_length(
-            self.bz_sigma_t[i],
-            self.bz_sigma_s[i, i],
-            self.bz_A,
+            self.bz_mat.sigma_t[i],
+            self.bz_mat.sigma_s[i, i],
+            self.bz_mat.avg_atomic_mass,
         )
         l_bz = np.sqrt(abs(l_bz_2))
         self.flux_escaped[i] = (
